requirements.txt
Flask
flask-cors
pydantic
python-dotenv
groq
